fastapi_app/requirements.txt
fastapi
numpy
requests
pillow
uvicorn